+++ checkpoint/orbax/checkpoint/_src/checkpointers/async_checkpointer.py
@@ -346,7 +346,12 @@ class AsyncCheckpointer(checkpointer.Checkpointer):
     )
 
   async def _save(
-      self, directory: epath.PathLike, *args, force: bool = False, **kwargs
+      self,
+      directory: epath.PathLike,
+      *args,
+      force: bool = False,
+      custom_metadata: dict[str, Any] | None = None,
+      **kwargs
   ):
     checkpoint_start_time = time.time()
     directory = epath.Path(directory)
@@ -402,7 +407,7 @@ class AsyncCheckpointer(checkpointer.Checkpointer):
       if utils.is_primary_host(self._primary_host):
         # Update StepMetadata after the handler save is complete.
         # (blocking write)
-        self._save_step_metadata(tmpdir.get(), custom_metadata=None)
+        self._save_step_metadata(tmpdir.get(), custom_metadata=custom_metadata)
       logging.info(
           '[process=%s][thread=%s] Async Save Callback [1/3]: Finalizing'
           ' Handler: %s on %s',
@@ -444,7 +449,12 @@ class AsyncCheckpointer(checkpointer.Checkpointer):
     )
 
   def save(
-      self, directory: epath.PathLike, *args, force: bool = False, **kwargs
+      self,
+      directory: epath.PathLike,
+      *args,
+      force: bool = False,
+      custom_metadata: dict[str, Any] | None = None,
+      **kwargs
   ):
     """Saves the given item to the provided directory.
 
@@ -460,13 +470,23 @@ class AsyncCheckpointer(checkpointer.Checkpointer):
       *args: additional args to provide to the CheckpointHandler's save method.
       force: if True, allows overwriting an existing directory. May add overhead
         due to the need to delete any existing files.
+      custom_metadata: a dictionary of custom metadata to be written to the
+        checkpoint directory via StepMetadata.
       **kwargs: additional keyword args to provide to the CheckpointHandler's
         save method.
 
     Raises:
       ValueError if the provided directory already exists.
     """
-    asyncio_utils.run_sync(self._save(directory, *args, force=force, **kwargs))
+    asyncio_utils.run_sync(
+        self._save(
+            directory,
+            *args,
+            force=force,
+            custom_metadata=custom_metadata,
+            **kwargs
+        )
+    )
 
   def restore(self, directory: epath.PathLike, *args, **kwargs) -> Any:
     """See superclass documentation."""